+++ packages/fetch-router/src/lib/controller.ts
@@ -1,7 +1,7 @@
 import type { Params, RoutePattern } from '@remix-run/route-pattern'
 
 import type { Middleware } from './middleware.ts'
-import type { RequestHandler } from './request-handler.ts'
+import type { RequestContext } from './request-context.ts'
 import type { RequestMethod } from './request-methods.ts'
 import type { Route, RouteMap } from './route-map.ts'
 
@@ -58,6 +58,19 @@ export type BuildAction<method extends RequestMethod | 'ANY', route extends stri
   route extends RoutePattern<infer pattern> ? Action<method, pattern> :
   route extends Route<infer _, infer pattern> ? Action<method, pattern> :
   never
+
+/**
+ * A request handler function that returns some kind of response.
+ *
+ * @param context The request context
+ * @return The response
+ */
+export interface RequestHandler<
+  method extends RequestMethod | 'ANY' = RequestMethod | 'ANY',
+  params extends Record<string, any> = {},
+> {
+  (context: RequestContext<method, params>): Response | Promise<Response>
+}
 
 /**
  * Runtime shape for a controller with middleware.